spare recalculation
On the aircraft on the way to my holiday and realised that I didn't check that the spare calculation happens only in a frame.  At the moment it would count two pins and if they add to ten then it would count as a spare but the problem is that the program needs to know that if the spare is in the frame.